reverted sqlite query builder and just changed using of adapter to structure

+++ src/Database/QueryBuilder/SqliteQueryBuilder.php
@@ -6,6 +6,7 @@ use Phoenix\Database\Adapter\AdapterInterface;
 use Phoenix\Database\Element\Column;
 use Phoenix\Database\Element\Index;
 use Phoenix\Database\Element\MigrationTable;
+use Phoenix\Exception\PhoenixException;
 
 class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterface
 {
@@ -101,13 +102,15 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
      */
     public function alterTable(MigrationTable $table)
     {
-        $queries = [];
-        $tmpTableName = '_' . $table->getName() . '_old_' . date('YmdHis');
-        $queries = array_merge($queries, $this->renameTable($table, $tmpTableName));
-        $queries = array_merge($queries, $this->createNewTable($table, $tmpTableName));
+        $queries = $this->addColumns($table);
+        if ($table->getColumnsToChange()) {
+            $tmpTableName = '_' . $table->getName() . '_old_' . date('YmdHis');
+            $queries = array_merge($queries, $this->renameTable($table, $tmpTableName));
+            $queries = array_merge($queries, $this->createNewTable($table, $tmpTableName));
 
-        $tableToDrop = new MigrationTable($tmpTableName);
-        $queries = array_merge($queries, $this->dropTable($tableToDrop));
+            $tableToDrop = new MigrationTable($tmpTableName);
+            $queries = array_merge($queries, $this->dropTable($tableToDrop));
+        }
 
         return $queries;
     }
@@ -156,28 +159,17 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
         return $query;
     }
 
-    private function dropIndex($indexName)
-    {
-        return 'DROP INDEX "' . $indexName . '"';
-    }
-
     private function createNewTable(MigrationTable $table, $tmpTableName)
     {
-        $queries = [];
+        if (is_null($this->adapter)) {
+            throw new PhoenixException('Missing adapter');
+        }
+        $oldColumns = $this->adapter->getStructure()->getTable($table->getName())->getColumns();
+        $columns = array_merge($oldColumns, $table->getColumnsToChange());
 
-        $oldTable = $this->adapter->getStructure()->getTable($table->getName());
-        $oldColumns = $oldTable->getColumns();
-        $columnsToDrop = $table->getColumnsToDrop();
-        $columnsToChange = $table->getColumnsToChange();
-
-        $columns = array_merge($oldColumns, $columnsToChange);
         $newTable = new MigrationTable($table->getName());
         $columnNames = [];
         foreach ($columns as $column) {
-            if (in_array($column->getName(), $columnsToDrop)) {
-                unset($oldColumns[$column->getName()]);
-                continue;
-            }
             $columnNames[] = $column->getName();
             if ($column->getSettings()->isAutoincrement()) {
                 $newTable->addPrimary($column);
@@ -185,30 +177,8 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
             }
             $newTable->addColumn($column->getName(), $column->getType(), $column->getSettings()->getSettings());
         }
-        foreach ($table->getColumns() as $newColumn) {
-            $newTable->addColumn($newColumn->getName(), $newColumn->getType(), $newColumn->getSettings()->getSettings());
-        }
 
-        $indexesToDrop = $table->getIndexesToDrop();
-        foreach ($oldTable->getIndexes() as $index) {
-            $queries[] = $this->dropIndex($index->getName());
-            if (in_array($index->getName(), $indexesToDrop)) {
-                continue;
-            }
-            $indexColumns = [];
-            foreach ($index->getColumns() as $indexColumn) {
-                $indexColumns[] = array_key_exists($indexColumn, $columnsToChange) ? $columnsToChange[$indexColumn]->getName() : $indexColumn;
-            }
-            $newTable->addIndex($indexColumns, $index->getType(), $index->getMethod(), $index->getName());
-        }
-
-        foreach ($table->getIndexes() as $index) {
-            $newTable->addIndex($index->getColumns(), $index->getType(), $index->getMethod(), $index->getName());
-        }
-
-        // foreign keys
-
-        $queries = array_merge($queries, $this->createTable($newTable));
+        $queries = $this->createTable($newTable);
         $queries[] = 'INSERT INTO ' . $this->escapeString($newTable->getName()) . ' (' . implode(',', $this->escapeArray($columnNames)) . ') SELECT ' . implode(',', $this->escapeArray(array_keys($oldColumns))) . ' FROM ' . $this->escapeString($tmpTableName);
         return $queries;
     }
